backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langgraph==0.2.60
langchain-openai==0.3.0
langchain-core==0.3.28
openai==1.58.1
pydantic==2.10.3
python-dotenv==1.0.1
sse-starlette==2.2.1
